Add metrics for total of errors
This adds a metrics counter for the total of errors in audito maldito.
The metric is called `audito_maldito_errors_total`. It has a single
label called `type` which currently only picks up errors for
journald read errors. The type is hardcoded to `journald_wait`.

The guidance is that we should only add errors to this metric that we
should alert on, this way we can write proper alertmanager rules for
this.
